Support manifest-beta.json

+++ publish-plugin.sh
@@ -4,9 +4,22 @@ PLUGIN_NAME=${PLUGIN_NAME:-${GITHUB_REPOSITORY##*/}}
 
 TAG_NAME=${GITHUB_REF##*/}
 MANIFEST_VERSION=$(jq -r .version manifest.json)
+MANIFEST_FILE=manifest.json
+
+# Check if beta manifest is newer: if so, use that for the release
+# (so betas can be made public later and have the right version number)
+#
+if [[ -f manifest-beta.json ]]; then
+    BETA_VERSION=$(jq -r .version manifest-beta.json)
+    if jq --arg v1 "$BETA_VERSION" --arg v2 "$MANIFEST_VERSION" -ne '($v1|split(".")) > ($v2|split("."))' >/dev/null; then
+        MANIFEST_VERSION="$BETA_VERSION"
+        MANIFEST_FILE=manifest-beta.json
+        cp manifest-beta.json manifest.json
+    fi
+fi
 
 if [[ "$MANIFEST_VERSION" != "$TAG_NAME" ]]; then
-    echo "ERROR: Commit is tagged '$TAG_NAME' but manifest version is '$MANIFEST_VERSION'"
+    echo "ERROR: Commit is tagged '$TAG_NAME' but $MANIFEST_FILE version is '$MANIFEST_VERSION'"
     exit 1
 fi
 
@@ -18,8 +31,8 @@ rm -f dist/main.css
 mkdir "${PLUGIN_NAME}"
 
 assets=()
-for f in main.js manifest*.json styles.css; do
-    if [[ -f dist/$f ]] && [[ ! -f $f ]]; then
+for f in main.js manifest.json styles.css; do
+    if [[ "$f" != manifest.json && -f "dist/$f" && ! -f "$f" ]]; then
         mv dist/$f $f;
     fi
     if [[ -f $f ]]; then
